+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Filters have permanent and ad-hoc relation targets (#113)
 - Filters can be registered to the cache to speed up query iteration (#114, #116, #122)
 - Relation targets can be specified by either component type or component index (#120)
+- `ExchangeX.Removes` can be called multiple times in chains (#124)
 
 ### Documentation